backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pydantic
pypdf
openai
python-dotenv
sentence-transformers
requests
